requirements_v23_windows.txt
# 솔로몬드 AI v2.3 Windows 호환 패키지 목록
# Windows 환경에서 안정적으로 작동하는 패키지들

# 웹 프레임워크
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# AI 및 음성 처리 (Windows 호환)
openai-whisper>=20231117
openai>=1.3.0
anthropic>=0.8.0

# 오디오 처리 (Windows 안정화)
soundfile>=0.12.1
librosa>=0.10.1
numpy>=1.24.0
scipy>=1.11.0

# 파일 처리
python-multipart>=0.0.6
aiofiles>=23.2.1

# 데이터 처리
pandas>=2.1.0
requests>=2.31.0

# 시스템 모니터링
psutil>=5.9.6

# 암호화 및 보안
cryptography>=41.0.0

# 환경 설정
python-dotenv>=1.0.0

# 로깅
loguru>=0.7.2

# AI 및 머신러닝 (핵심)
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
easyocr>=1.7.0

# 화자 분리 (Speaker Diarization)
pyannote.audio>=3.1.0
speechbrain>=0.5.0

# 이미지 처리 (선택적)
Pillow>=10.1.0

# 웹 클라이언트
httpx>=0.25.0

# 타임존 처리
pytz>=2023.3

# JSON 처리
pydantic>=2.5.0

# 템플릿 엔진
jinja2>=3.1.2

# 비동기 처리
asyncio-timeout>=4.0.3

# 개발 도구 (선택적)
black>=23.11.0
isort>=5.12.0
